Add RFC for low-level timseries write API (#16)
This RFC proposes a low-level `TimeSeries` API which is analogous to SlateDb's `Db` and OpenData's `Log`. The idea is to have a simple, permissive low-level API with all the complexity from systems like OTEL and Prometheus layered on top. We started with the write API because that is relatively simple compared to the read side.